requirements.txt
celery==4.3
minio==4.0.18
newspaper3k==0.2.8
gevent==1.4.0

